WIP: https://github.com/valdisiljuconoks/LocalizationProvider/issues/203 * moved GetTranslationHandler
