+++ src/DbLocalizationProvider.Storage.PostgreSQL/ConfigurationContextExtensions.cs
@@ -4,7 +4,6 @@
 using System;
 using DbLocalizationProvider.Abstractions;
 using DbLocalizationProvider.Commands;
-using DbLocalizationProvider.Queries;
 using DbLocalizationProvider.Sync;
 
 namespace DbLocalizationProvider.Storage.PostgreSql
@@ -36,8 +35,6 @@ namespace DbLocalizationProvider.Storage.PostgreSql
 
             context.TypeFactory.ForQuery<UpdateSchema.Command>().SetHandler<SchemaUpdater>();
             context.TypeFactory.ForQuery<SyncResources.Query>().SetHandler<ResourceSynchronizer>();
-
-            context.TypeFactory.ForQuery<GetTranslation.Query>().SetHandler<GetTranslationHandler>();
 
             context.TypeFactory.ForCommand<RemoveTranslation.Command>().SetHandler<RemoveTranslationHandler>();
